Added auto-completion, tagging, src block code snippet capture
Turns out it's rather trivial to add some auto-complete values to the
prompt supplied by `%^{}`! and free tagging to boot.

Still do have a bug where it tries to expand the content in `%i` if it
has org-capture template expansions, which is livable if I don't have
to capture org-capture-like code 🌚.