hook to display images after org-babel runs and pl tempo :file add
Added a hook to display images in org after babel runs so our plantUml
diagrams will update as we edit them. This'll be running every time we
execute org-babel until we figure out how to extract the language from
the hook and _only_ run the command if we're doing a plantUml thing.

Also added a `:file diagram.png` by default to the PlantUml structure
template for convenience.