Formatting tweaks to leader keys and newline for load languages
Org's being weird about the link after load languages and displaying
it besides the top-level header when blocks are collapsed. Seeing if
adding a newline to it fixes the problem.

Put 'org-roam-node-insert-immediate' besides 'org-roam-node-insert'
since both their changes result in the same changes to the actual
state of the roam database, although the order of declaration is
meaningless here.