Emacs 30.1 update, remove ox-hugo
Big change is that `dired-single` needs to be downloaded directly, I'm
guessing it was either removed from the melpa repo OR I was just
always doing this and forgot. Either way, I added
`~/.emacs.d/mis-elisp/` to the load path and added instructions that
the script should be downloaded to it.

Removed ox-hugo since I never use it. I probably should start culling
packages in order to clean up the config a bit.

`ob-go` was also throwing an error until I placed it before the
initialization of `go-mode`, but I didn't commit to the config because
I couldn't make sense of why. If I start doing more golang
development, I'll dig deeper.